Prepare package for publication
