+++ test/bubble_sort.spec.ts
@@ -1,37 +1,36 @@
-import {functionalQuickSort, QuickSort} from './quick_sort'
-import * as expect from 'expect'
-import {comparator} from "./comparator";
-import {functionalBubbleSort} from "./bubble_sort";
+import { BubbleSort, functionalBubbleSort } from '../src/bubble_sort'
+import expect from 'expect'
+import { comparator } from '../src/comparator'
 
-describe('src/quick_sort.ts', () => {
-    describe('QuickSort.sort()', () => {
-        it('should sort numbers in asceding order using quick_sort algorithm', () => {
+describe('src/bubble_sort.ts', () => {
+    describe('BubbleSort.sort()', () => {
+        it('should sort numbers in asceding order using bubble sort algorithm', () => {
             let input: Array<number> = [1, 4, 5, 2, 6]
-            const bubbleSort = new QuickSort()
+            const bubbleSort = new BubbleSort()
             let result: Array<number> = bubbleSort.sort(input)
             expect(result).toEqual([1, 2, 4, 5, 6])
         })
 
-        it('should sort letters in asceding order using quick_sort algorithm', () => {
+        it('should sort letters in asceding order using bubble sort algorithm', () => {
             let input: Array<string> = ['c', 'b', 'd', 'a', 'e']
-            const bubbleSort = new QuickSort()
+            const bubbleSort = new BubbleSort()
             let result: Array<string> = bubbleSort.sort(input)
             expect(result).toEqual(['a', 'b', 'c', 'd', 'e'])
         })
 
         it('should sort arrays in asceding order', () => {
             let input: Array<Array<number>> = [[1, 2, 3], [1, 3, 6, 4], [1]]
-            const bubbleSort = new QuickSort()
+            const bubbleSort = new BubbleSort()
             let result: Array<Array<number>> = bubbleSort.sort(input)
             expect(result).toEqual([[1], [1, 2, 3], [1, 3, 6, 4]])
         })
     })
 
-    describe('functionalQuickSort', () => {
+    describe('functionalBubbleSort', () => {
         describe('()', () => {
             it('should sort numbers in ascending order', () => {
                 let input: Array<number> = [1, 4, 5, 2, 6]
-                let result: Array<number> = functionalQuickSort(
+                let result: Array<number> = functionalBubbleSort(
                     input,
                     comparator
                 )
