+++ tsconfig.json
@@ -1,12 +1,17 @@
 {
     "compilerOptions": {
+        "target": "ESNext",
         "module": "CommonJS",
+        "lib": ["ESNext", "ES6", "ES2020", "ES2019", "DOM"],
+        "outDir": "dist",
+        "declaration": true,
+        "strict": true,
+        "esModuleInterop": true,
         "noImplicitAny": true,
         "removeComments": true,
         "preserveConstEnums": true,
-        "outFile": "../../built/local/tsc.js",
         "sourceMap": true
     },
     "include": ["src/**/*"],
-    "exclude": ["node_modules", "**/*.spec.ts"]
+    "exclude": ["node_modules", "dist", "test"]
 }
